this commit rearranges file and folder locations to how it should be going forward, before pulling in the changes from the coordinate system update.
