--- py4DSTEM/process/rdf/__init__.py
@@ -1,1 +1,0 @@
-from .rdf import *
